added roomwide floor collider and changed teleport to on trigger
scripts changed:
Teleport.cs
BasicEnemyAI.cs
BasicBarryAI.cs

Changed floor colliders to be one large collider that spans the whole
floor. This was to stop the enemies getting stuck on the tile edges.
Changed Teleport to be on trigger instead of collide so it can trigger
from the tiles below the collider.